# Added Errorstats info section to enable keeping track of the different errors that occur within Redis; Added failed_calls to commandstats