--- a/src/networking.c
+++ b/src/networking.c
@@ -411,15 +411,15 @@ void afterErrorReply(client *c, const char *s, size_t len) {
      * If the string already starts with "-..." then the error prefix
      * is provided by the caller ( we limit the search to 32 chars). Otherwise we use "-ERR". */
     if (s[0] != '-') {
-        incrementError("ERR", 3);
+        incrementErrorCount("ERR", 3);
     } else {
         char* spaceloc = memchr(s, ' ', len < 32 ? len : 32);
         if (spaceloc) {
             const size_t errEndPos = (size_t)(spaceloc - s);
-            incrementError(s+1, errEndPos-1);
+            incrementErrorCount(s+1, errEndPos-1);
         } else {
             /* Fallback to ERR if we can't retrieve the error prefix */
-            incrementError("ERR", 3);
+            incrementErrorCount("ERR", 3);
         }
     }