--- a/src/server.h
+++ b/src/server.h
@@ -2140,7 +2140,7 @@ void populateCommandTable(void);
 void resetCommandTableStats(void);
 void resetErrorTableStats(void);
 void adjustOpenFilesLimit(void);
-void incrementError(const char* fullerr, size_t namelen);
+void incrementErrorCount(const char* fullerr, size_t namelen);
 void closeListeningSockets(int unlink_unix_socket);
 void updateCachedTime(int update_daylight_info);
 void resetServerStats(void);