--- a/tests/cluster/tests/18-info.tcl
+++ b/tests/cluster/tests/18-info.tcl
@@ -3,7 +3,7 @@
 source "../tests/includes/init-tests.tcl"
 
 test "Create a primary with a replica" {
-    create_cluster 2 2
+    create_cluster 2 0
 }
 
 test "Cluster should start ok" {
@@ -12,8 +12,6 @@ test "Cluster should start ok" {
 
 set primary1 [Rn 0]
 set primary2 [Rn 1]
-set replica1 [Rn 2]
-set replica2 [Rn 3]
 
 proc cmdstat {instace cmd} {
     return [cmdrstat $cmd $instace]
@@ -28,19 +26,11 @@ test "errorstats: rejected call due to MOVED Redirection" {
     $primary2 config resetstat
     assert_match {} [errorstat $primary1 MOVED]
     assert_match {} [errorstat $primary2 MOVED]
-    # we know that one of the commands will have a MOVED reply
-    catch {$primary1 set a b} replyP1
-    catch {$primary2 set a b} replyP2
-    if { [string range $replyP1 0 4] eq {MOVED} } {
-        assert_match {*count=1*} [errorstat $primary1 MOVED]
-        assert_match {*calls=0,*,rejected_calls=1,failed_calls=0} [cmdstat $primary1 set]
-        assert_match {OK} $replyP2
-        assert_match {} [errorstat $primary2 MOVED]
-    }
-    if { [string range $replyP2 0 4] eq {MOVED} } {
-        assert_match {*count=1*} [errorstat $primary2 MOVED]
-        assert_match {*calls=0,*,rejected_calls=1,failed_calls=0} [cmdstat $primary2 set]
-        assert_match {OK} $replyP1
-        assert_match {} [errorstat $primary1 MOVED]
-    }
+    # we know that the primary 2 will have a MOVED reply
+    catch {$primary1 set key{0x0000} b} replyP1
+    catch {$primary2 set key{0x0000} b} replyP2
+    assert_match {OK} $replyP1
+    assert_match {} [errorstat $primary1 MOVED]
+    assert_match {*count=1*} [errorstat $primary2 MOVED]
+    assert_match {*calls=0,*,rejected_calls=1,failed_calls=0} [cmdstat $primary2 set]
 }