… tests/unit/info.tcl