this PR pushes forward the observability on overall error statistics and command statistics within redis-server:

- It extends `INFO COMMANDSTATS` to have
  - failed_calls in - so we can keep track of errors that happen from the command itself, broken by command.
  - rejected_calls - so we can keep track of errors that were triggered outside the commmand processing per se ( example of wrong arity, oom, etc... )
Sample ( notice rejected_calls ):
  ```
  127.0.0.1:6379> set k 
  (error) ERR wrong number of arguments for 'set' command
  127.0.0.1:6379> info commandstats
  # Commandstats
  cmdstat_set:calls=0,usec=0,usec_per_call=0.00,rejected_calls=1,failed_calls=0
  cmdstat_info:calls=8,usec=1077,usec_per_call=134.62,rejected_calls=0,failed_calls=0
  ```

- Adds a new section to INFO, named `ERRORSTATS` that enables keeping track of the different errors that occur within redis ( within processCommand and call ) based uppong the reply Error Prefix ( The first word after the "-", up to the first space ). 
  
  Sample:
  ```
  127.0.0.1:6379> auth a
  (error) ERR AUTH <password> called without any password configured for the default user. Are you sure your configuration is correct?
  127.0.0.1:6379> set k
  (error) ERR wrong number of arguments for 'set' command
  127.0.0.1:6379> ACL SETUSER alice on >p1pp0 ~cached:* +get +info
  OK
  127.0.0.1:6379> auth alice p1pp0
  OK
  127.0.0.1:6379> set a b
  (error) NOPERM this user has no permissions to run the 'set' command or its subcommand
  127.0.0.1:6379> info errorstats
  # Errorstats
  errorstat_ERR:count=2
  errorstat_NOPERM:count=1
  ```


- This PR also fixes RM_ReplyWithError so that it can be correctly identified as an error reply.

---
To test the added features:
```
tclsh tests/test_helper.tcl --verbose --single unit/info
tclsh tests/cluster/run.tcl --single tests/18-info.tcl
```

The added tests are divided as follow:
- **failed** call authentication error
- **failed** call NOGROUP error
- **rejected** call due to wrong arity
- **rejected** call by OOM error
- **rejected** call by authorization error
- **rejected** call rejected due to MOVED Redirection