@madolson 2020-12-21
Also related to Oran's top level comment, we don't always use "addReplyError*" for all of the errors, we need to go through and audit them.