@filipecosta90 2020-12-22
> do we intend to cover only the first group? or also the second?
I believe we can see value out of both groups right? meaning the more errors we can keep track without paying a penalty the better.

> If we intend to extend this feature to match other types of errors, it's probably wrong to do an if-else chain of calls to `strncmp`, and then a lookup in a dict (or an enum lookup into an array).
> maybe we should have a rax in which we can search by prefix, and implicitly support all type of errors.
> i think it would be nice to examine this feature after adding a few more errors as an example (like OOM, WRONGTYPE, NOSCRIPT, LOADING, BUSY, BUSYKEY.

@oranagra @madolson, following what you've mentioned I've updated the code to use the Error Prefix ( The first word after the "-", up to the first space ) as the rax key (lowercased). In that manner we have the desired implicit support for all errors. Can you guys check the new commit. 

Sample output of errorstats now:
```
127.0.0.1:6379> auth a
(error) ERR AUTH <password> called without any password configured for the default user. Are you sure your configuration is correct?
127.0.0.1:6379> set k
(error) ERR wrong number of arguments for 'set' command
127.0.0.1:6379> info errorstats
# Errorstats
errorstat_auth:count=1
errorstat_wrong:count=1
```