@oranagra 2020-12-22
@filipecosta90 looking at the example in your last post (didn't read the code yet), i think these two cases should fall into the standard `err` since the error code is `-ERR` ("wrong" and "AUTH" are just part of the text).  
i.e.
```
errorstat_err:count=2
``` 
not
```
errorstat_auth:count=1
errorstat_wrong:count=1
```

if we think it is important for redis to distinguish between them, then it's also important for client libraries, and we should change the code and promote them to have their own error code.