@filipecosta90 2020-12-22
> we do not want to **optionally** remove the spurious dash, this is in fact an indication that this method (`afterErrorReply`) is called in the wrong place, and given the wrong input (incomplete error message).
> if we do that, the dash will always be present (and we can remove it, and even assert that it's there).

@oranagra agree. In the latest code push I assume that if no `-` is present than the error is the default `ERR` (same as done in `addReplyErrorLength`), otherwise we use the first after the `-` until the first space. 

I've also removed the error name from the error struct and I'm using the error prefix as is ( meaning no lowercase ). 

I believe all comments have been address. wdyt?