@oranagra 2020-12-23
@filipecosta90 it looks like you added a bunch of commits, and a rebase/merge from unstable, and then squashed it all, and i can't seem to be able to review the changes easily.
please avoid it in the future.

i'm ok with amending a commit and force pushing it (github let's me view the diffs), or to add more and more commits which will then be squashed when the PR is merged.
rebase from unstable or squashing your commits should be done when no one is still looking at the PR, and if you have to grab fresh code from unstable, do it with merge and no squashing, so i can skip that commit.