@oranagra 2020-12-23
@madolson i agree. i'm still editing my CR post when i'm suggesting a better alternative.