@oranagra 2020-12-27
@redis/core-team please approve new info fields:
a per command stat `rejected_calls` and `failed_calls`
a per-error type counters (new info section)
and a new stat for total error count.
details at the top.

@filipecosta90 can you please make a redis-doc PR? thank you!