@filipecosta90 2020-12-27
> @filipecosta90 can you please make a redis-doc PR? thank you!

@oranagra @itamarhaber with regards to the redis-doc PR can you guys check https://github.com/redis/redis-doc/pull/1472 ?