@filipecosta90 2020-12-29
> My approval is for the new API syntax, I would still like to fix some of the minor comments.

@madolson @oranagra addressed the comments in the new commit.