@filipecosta90 2020-12-31
> That said, I doubt that performance is really an issue here (@filipecosta90 you can probably verify that in a breeze). Everything else is a matter of personal taste and I don't feel that strongly about it - so +1.

@oranagra @yossigo @itamarhaber @madolson  I was getting the numbers for the standalone and cluster variations and noticed the following:
- performance is not affected here as you can check bellow
- one thing I believe we need to fix prior merge is that MOVED/CLUSTERDOWN/CROSSSLOT/... ( all error replies caused within `clusterRedirectClient()` are being accounted on the errorstats section, but not on the commandstats ( given we're not incrementing the rejected_calls there ). I believe we should... . If you agree tomorrow I'll fix it and include tests to verify it...

As a quick note @oranagra , to test the OSS Cluster variation I needed to update locally the branch to include the already merged changes by #8223.

----- 
## standalone numbers of an OOO error test

Used command and error replies:
```
$ redis-benchmark --threads 4 -h 10.3.0.110 -e -n 10000000 -t set
(...)
(...)
Error from server: OOM command not allowed when used memory > 'maxmemory'.
Error from server: OOM command not allowed when used memory > 'maxmemory'.
Error from server: OOM command not allowed when used memory > 'maxmemory'.
(...)
(...)
```

Without errorstats:
```
Summary:
  throughput summary: 204611.95 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
        0.225     0.048     0.223     0.303     0.335     1.487
```

With errorstats:
```
Summary:
  throughput summary: 198503.30 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
        0.232     0.048     0.231     0.311     0.343     3.319

$ $ redis-cli -h 10.3.0.110
10.3.0.110:6379> info errorstats
# Errorstats
errorstat_OOM:count=10000000
10.3.0.110:6379> info commandstats
# Commandstats
cmdstat_info:calls=2,usec=61,usec_per_call=30.50,rejected_calls=0,failed_calls=0
cmdstat_command:calls=2,usec=801,usec_per_call=400.50,rejected_calls=0,failed_calls=0
cmdstat_config:calls=5,usec=56,usec_per_call=11.20,rejected_calls=0,failed_calls=0
cmdstat_set:calls=348459,usec=106708,usec_per_call=0.31,rejected_calls=10000000,failed_calls=0
cmdstat_flushall:calls=1,usec=20,usec_per_call=20.00,rejected_calls=0,failed_calls=0
```


----- 
## OSS cluster numbers of an MOVED errors worst case scenario ( all moved )

Used command and error replies:
```
$ redis-benchmark --threads 4 -h 10.3.0.110 -p 30001 -e -n 10000000 -t set
(...)
(...)
Error from server: MOVED 13782 127.0.0.1:300038 (overall: 0.261).261)
Error from server: MOVED 13782 127.0.0.1:300030.407 (overall: 0.257))
Error from server: MOVED 13782 127.0.0.1:300032 (overall: 0.256).256)
(...)
(...)
```

Without errorstats:
```
Summary:
  throughput summary: 183874.23 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
        0.252     0.048     0.255     0.327     0.359    18.207
```

With errorstats ( notice errorstats is populated with info but not commandstats ):
```
Summary:
  throughput summary: 184723.38 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
        0.250     0.048     0.247     0.327     0.351    14.719

ubuntu@ip-10-3-0-195:~$ redis-cli -h 10.3.0.110 -p 30001 
10.3.0.110:30001> info errorstats
# Errorstats
errorstat_MOVED:count=10000000
10.3.0.110:30001> info commandstats
# Commandstats
cmdstat_psync:calls=1,usec=314,usec_per_call=314.00,rejected_calls=0,failed_calls=0
cmdstat_command:calls=1,usec=414,usec_per_call=414.00,rejected_calls=0,failed_calls=0
cmdstat_info:calls=4,usec=135,usec_per_call=33.75,rejected_calls=0,failed_calls=0
cmdstat_ping:calls=1,usec=0,usec_per_call=0.00,rejected_calls=0,failed_calls=0
cmdstat_cluster:calls=8,usec=1059,usec_per_call=132.38,rejected_calls=0,failed_calls=0
cmdstat_replconf:calls=204,usec=129,usec_per_call=0.63,rejected_calls=0,failed_calls=0
cmdstat_config:calls=2,usec=32,usec_per_call=16.00,rejected_calls=0,failed_calls=0

```